# specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Biotech M&A Predictor API
  version: 1.0.0
  description: |
    API for accessing biotech M&A predictions, company intelligence, and market signals.

    This API serves institutional clients (hedge funds, PE firms, investment banks) with:
    - Real-time M&A probability scores and rankings
    - Comprehensive company profiles and drug pipeline data
    - Historical signal tracking and analysis
    - Automated report generation and delivery
    - Configurable alerts for market events

    ## Authentication
    All requests require an API key passed in the `X-API-Key` header.

    ## Rate Limiting
    - Standard tier: 100 requests/minute
    - Premium tier: 500 requests/minute
    - Enterprise tier: Custom limits

    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the window resets

  contact:
    name: API Support
    email: api-support@biotechma.io
  license:
    name: Proprietary
servers:
  - url: https://api.biotechma.io/v1
    description: Production server
  - url: https://api-staging.biotechma.io/v1
    description: Staging server
security:
  - ApiKeyAuth: []
...
paths:
...
  /predictions/watchlist:
    get:
      tags:
        - Predictions
      summary: Get current M&A watchlist
      description: |
        Retrieve the current M&A watchlist with companies ranked by acquisition probability.
        This is the primary endpoint for identifying high-probability acquisition targets.
      operationId: getWatchlist
      parameters:
        - name: min_score
          in: query
          description: Minimum M&A score threshold
          schema:
            type: number
            format: float
            default: 60.0
            minimum: 0
            maximum: 100
        - name: limit
          in: query
          description: Number of companies to return
          schema:
            type: integer
            default: 25
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  generated_at:
                    type: string
                    format: date-time
                  watchlist:
                    type: array
                    items:
                      $ref: '#/components/schemas/WatchlistEntry'
              example:
                generated_at: "2024-01-15T15:00:00Z"
                watchlist:
                  - rank: 1
                    ticker: "ABCD"
                    company_name: "ABC Therapeutics Inc."
                    ma_score: 78.3
                    ma_score_change_7d: 5.2
                    acquisition_probability: "High"
                    market_cap: 450.5
                    key_catalysts:
                      - "Cash runway below 9 months"
                      - "CFO departure"
                      - "Phase 2 trial success in lead program"
                    predicted_timeline: "3-6 months"
                    top_potential_acquirers:
                      - ticker: "BIGP"
                        company_name: "Big Pharma Corp"
                        match_score: 85.2
                      - ticker: "MEGA"
                        company_name: "Mega Biopharma"
                        match_score: 78.9
                  - rank: 2
                    ticker: "XYZT"
                    company_name: "XYZ Biopharma"
                    ma_score: 72.1
                    ma_score_change_7d: -2.1
                    acquisition_probability: "High"
                    market_cap: 320.8
                    key_catalysts:
                      - "Recent analyst downgrades"
                      - "Regulatory setback in EU"
                    predicted_timeline: "6-12 months"
                    top_potential_acquirers:
                      - ticker: "GIANT"
                        company_name: "Giant Pharma Inc"
                        match_score: 82.1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /predictions/top:
    get:
      tags:
        - Predictions
      summary: Get top N acquisition candidates
      description: |
        Retrieve the top acquisition candidates with detailed scoring and rationale.
        Similar to watchlist but with more detailed analysis.
      operationId: getTopCandidates
      parameters:
        - name: limit
          in: query
          description: Number of top candidates to return
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
        - name: therapeutic_area
          in: query
          description: Filter by therapeutic area
          schema:
            type: string
        - name: max_market_cap
          in: query
          description: Maximum market cap in millions USD
          schema:
            type: number
            format: float
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  generated_at:
                    type: string
                    format: date-time
                  candidates:
                    type: array
                    items:
                      $ref: '#/components/schemas/AcquisitionCandidate'
              example:
                generated_at: "2024-01-15T15:00:00Z"
                candidates:
                  - ticker: "ABCD"
                    company_name: "ABC Therapeutics Inc."
                    ma_score: 78.3
                    market_cap: 450.5
                    scoring_breakdown:
                      financial_stress: 35.0
                      pipeline_value: 25.5
                      market_dynamics: 10.3
                      corporate_activity: 7.5
                    acquisition_rationale: "Strong phase 2 data in oncology, but limited cash runway creates urgency for strategic transaction"
                    strategic_fit:
                      - acquirer_profile: "Large cap pharma seeking oncology expansion"
                        fit_score: 88.5
                      - acquirer_profile: "Mid-cap biotech with commercial infrastructure"
                        fit_score: 72.3
                    valuation_estimate:
                      base_case: 650.0
                      bull_case: 850.0
                      bear_case: 500.0
                      premium_to_current: 44.4
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /predictions/{ticker}/acquirers:
    get:
      tags:
        - Predictions
      summary: Get potential acquirers for target
      description: |
        For a given target company, retrieve a ranked list of potential acquirers
        with strategic fit scores and acquisition rationale.
      operationId: getPotentialAcquirers
      parameters:
        - name: ticker
          in: path
          required: true
          description: Target company ticker symbol
          schema:
            type: string
            example: "ABCD"
        - name: limit
          in: query
          description: Number of potential acquirers to return
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  target:
                    type: object
                    properties:
                      ticker:
                        type: string
                      company_name:
                        type: string
                      market_cap:
                        type: number
                      ma_score:
                        type: number
                  potential_acquirers:
                    type: array
                    items:
                      $ref: '#/components/schemas/PotentialAcquirer'
                  generated_at:
                    type: string
                    format: date-time
              example:
                target:
                  ticker: "ABCD"
                  company_name: "ABC Therapeutics Inc."
                  market_cap: 450.5
                  ma_score: 78.3
                potential_acquirers:
                  - rank: 1
                    ticker: "BIGP"
                    company_name: "Big Pharma Corp"
                    match_score: 85.2
                    market_cap: 125000.0
                    strategic_rationale: "Portfolio gap in NSCLC, recent divestiture generated $3B in capital, strong track record in oncology M&A"
                    fit_factors:
                      therapeutic_alignment: 95.0
                      pipeline_complementarity: 88.0
                      financial_capacity: 98.0
                      geographic_synergy: 75.0
                      cultural_fit: 72.0
                    estimated_offer_range:
                      low: 600.0
                      high: 900.0
                      premium_percentage: "33-100%"
                    acquisition_likelihood: "High"
                    historical_deal_activity:
                      deals_last_5y: 8
                      avg_deal_size: 850.0
                      therapeutic_focus: ["oncology", "immunology"]
                  - rank: 2
                    ticker: "MEGA"
                    company_name: "Mega Biopharma"
                    match_score: 78.9
                    market_cap: 45000.0
                    strategic_rationale: "Expanding oncology franchise, ABC-101 complements existing checkpoint inhibitor portfolio"
                    fit_factors:
                      therapeutic_alignment: 92.0
                      pipeline_complementarity: 85.0
                      financial_capacity: 82.0
                      geographic_synergy: 68.0
                      cultural_fit: 80.0
                    estimated_offer_range:
                      low: 550.0
                      high: 750.0
                      premium_percentage: "22-67%"
                    acquisition_likelihood: "Medium-High"
                    historical_deal_activity:
                      deals_last_5y: 4
                      avg_deal_size: 420.0
                      therapeutic_focus: ["oncology", "rare disease"]
                generated_at: "2024-01-15T15:00:00Z"
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /predictions/matches:
    get:
      tags:
        - Predictions
      summary: Get all target-acquirer pairings
      description: |
        Retrieve all potential target-acquirer pairings with match scores above threshold.
        Useful for market-wide M&A opportunity analysis.
      operationId: getMatches
      parameters:
        - name: min_match_score
          in: query
          description: Minimum match score threshold
          schema:
            type: number
            format: float
            default: 70.0
            minimum: 0
            maximum: 100
        - name: target_ticker
          in: query
          description: Filter by specific target company
          schema:
            type: string
        - name: acquirer_ticker
          in: query
          description: Filter by specific acquirer company
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Successful response
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TargetAcquirerMatch'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  generated_at:
                    type: string
                    format: date-time
              example:
                data:
                  - match_id: "match_001"
                    target:
                      ticker: "ABCD"
                      company_name: "ABC Therapeutics Inc."
                      market_cap: 450.5
                      ma_score: 78.3
                    acquirer:
                      ticker: "BIGP"
                      company_name: "Big Pharma Corp"
                      market_cap: 125000.0
                    match_score: 85.2
                    deal_probability: "High"
                    estimated_timeline: "3-6 months"
                    strategic_thesis: "BIGP seeking to fill NSCLC gap after patent expiration, ABC-101 Phase 2 data de-risks asset"
                    estimated_valuation: 750.0
                    premium_percentage: 66.7
                pagination:
                  page: 1
                  limit: 50
                  total_results: 156
                  total_pages: 4
                generated_at: "2024-01-15T15:00:00Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication. Contact support to obtain your key.
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current time window
      schema:
        type: integer
        example: 100
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current time window
      schema:
        type: integer
        example: 95
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit window resets
      schema:
        type: integer
        example: 1705329600
  schemas:
...
    WatchlistEntry:
      type: object
      properties:
        rank:
          type: integer
          description: Ranking on watchlist
          example: 1
        ticker:
          type: string
          example: "ABCD"
        company_name:
          type: string
          example: "ABC Therapeutics Inc."
        ma_score:
          type: number
          format: float
          example: 78.3
        ma_score_change_7d:
          type: number
          format: float
          example: 5.2
        acquisition_probability:
          type: string
          enum: [Low, Medium, Medium-High, High, Critical]
          example: "High"
        market_cap:
          type: number
          format: float
          example: 450.5
        key_catalysts:
          type: array
          items:
            type: string
          example: ["Cash runway below 9 months", "CFO departure"]
        predicted_timeline:
          type: string
          description: Estimated timeline for potential acquisition
          example: "3-6 months"
        top_potential_acquirers:
          type: array
          items:
            type: object
            properties:
              ticker:
                type: string
              company_name:
                type: string
              match_score:
                type: number
                format: float
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Results per page
          example: 50
        total_results:
          type: integer
          description: Total number of results
          example: 234
        total_pages:
          type: integer
          description: Total number of pages
          example: 5
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "invalid_request"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid ticker symbol provided"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
        request_id:
          type: string
          description: Unique request identifier for support
          example: "req_abc123xyz"
  responses:
...
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "unauthorized"
            message: "Invalid API key provided"
            request_id: "req_abc123xyz"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Company with ticker 'INVALID' not found"
            request_id: "req_abc123xyz"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "rate_limit_exceeded"
            message: "Rate limit exceeded. Limit resets in 45 seconds."
            details:
              limit: 100
              window: "1 minute"
              reset_at: 1705329600
            request_id: "req_abc123xyz"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "internal_error"
            message: "An unexpected error occurred. Please try again later."
            request_id: "req_abc123xyz"